Fix creating, updating, finding and deleting funeral service methods
Fix creating, updating, finding and deleting funeral service methods

+++ reservation_module/src/main/java/reservation_module/services/FuneralServiceImpl.java
@@ -2,17 +2,25 @@ package reservation_module.services;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.HttpStatus;
 import reservation_module.api.v1.mapper.FuneralMapper;
 import reservation_module.api.v1.mapper.GraveMapper;
+import reservation_module.api.v1.model.DeceasedDTO;
 import reservation_module.api.v1.model.FuneralDTO;
 import reservation_module.api.v1.model.FuneralListDTO;
 import reservation_module.exceptions.ResourceNotFoundException;
+import reservation_module.models.Deceased;
 import reservation_module.models.Funeral;
 import reservation_module.models.Grave;
+import reservation_module.repositories.DeceasedRepository;
 import reservation_module.repositories.FuneralRepository;
+import reservation_module.repositories.GraveRepository;
 
 import javax.transaction.Transactional;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -22,12 +30,16 @@ public class FuneralServiceImpl implements FuneralService {
 
     private final FuneralRepository funeralRepository;
     private final FuneralMapper funeralMapper;
+    private final GraveRepository graveRepository;
     private final GraveMapper graveMapper;
+    private final DeceasedRepository deceasedRepository;
 
-    public FuneralServiceImpl(FuneralRepository funeralRepository, FuneralMapper funeralMapper, GraveMapper graveMapper) {
+    public FuneralServiceImpl(FuneralRepository funeralRepository, FuneralMapper funeralMapper, GraveRepository graveRepository, GraveMapper graveMapper, DeceasedRepository deceasedRepository) {
         this.funeralRepository = funeralRepository;
         this.funeralMapper = funeralMapper;
+        this.graveRepository = graveRepository;
         this.graveMapper = graveMapper;
+        this.deceasedRepository = deceasedRepository;
     }
 
 
@@ -40,6 +52,11 @@ public class FuneralServiceImpl implements FuneralService {
                     FuneralDTO funeralDTO = funeralMapper.funeralToFuneralDTO(funeral);
                     if (funeralDTO.getGrave() != null) {
                         funeralDTO.getGrave().setFuneralId(funeralDTO.getId());
+                        if (funeralDTO.getGrave().getDeceased() != null) {
+                            for (DeceasedDTO deceasedDTO: funeralDTO.getGrave().getDeceased()) {
+                                deceasedDTO.setGraveId(funeralDTO.getGrave().getId());
+                            }
+                        }
                     }
                     return funeralDTO;
                 })
@@ -48,20 +65,25 @@ public class FuneralServiceImpl implements FuneralService {
     }
 
     @Override
-    public FuneralDTO getFuneralById(Long funeralId) {
+    public ResponseEntity getFuneralById(Long funeralId) {
         Optional<Funeral> funeralOptional = funeralRepository.findById(funeralId);
 
-        if(!funeralOptional.isPresent()) {
-            log.error("Funeral not found for ID: " + funeralId.toString());
+        if (!funeralOptional.isPresent()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Funeral not found for id: " + funeralId.toString());
         }
 
         FuneralDTO funeralDTO = funeralMapper.funeralToFuneralDTO(funeralOptional.get());
 
         if (funeralOptional.get().getGrave() != null) {
             funeralDTO.getGrave().setFuneralId(funeralDTO.getId());
+            if (funeralDTO.getGrave().getDeceased() != null) {
+                for (DeceasedDTO deceasedDTO: funeralDTO.getGrave().getDeceased()) {
+                    deceasedDTO.setGraveId(funeralDTO.getGrave().getId());
+                }
+            }
         }
 
-        return funeralDTO;
+        return ResponseEntity.status(HttpStatus.OK).body(funeralDTO);
     }
 
     @Override
@@ -70,12 +92,42 @@ public class FuneralServiceImpl implements FuneralService {
     }
 
     @Override
-    @Transactional
-    public FuneralDTO saveFuneralByDTO(Long id, FuneralDTO funeralDTO) {
-        Funeral funeral = funeralMapper.funeralDTOToFuneral(funeralDTO);
-        funeral.setId(id);
+    public ResponseEntity saveFuneralByDTO(FuneralDTO funeralDTO) {
+        Optional<Funeral> funeralOptional = funeralRepository.findById(funeralDTO.getId());
 
-        return saveAndReturnDTO(funeral);
+        if (!funeralOptional.isPresent()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Funeral not found for id: " + funeralDTO.getId().toString());
+        }
+
+        Funeral funeral = funeralMapper.funeralDTOToFuneral(funeralDTO);
+
+        if (funeralDTO.getReservationDate() != null) {
+            funeral.setReservationDate(funeralDTO.getReservationDate());
+        }
+        if (funeralDTO.getDate() != null) {
+            funeral.setDate(funeralDTO.getDate());
+        }
+        if (funeralDTO.getFuneralDirectorId() != null) {
+            funeral.setFuneralDirectorId(funeralDTO.getFuneralDirectorId());
+        }
+        if (funeralDTO.getUserId() != null) {
+            funeral.setUserId(funeralDTO.getUserId());
+        }
+
+        funeralRepository.save(funeral);
+
+        FuneralDTO returnDTO = funeralMapper.funeralToFuneralDTO(funeral);
+
+        if (returnDTO.getGrave() != null) {
+            returnDTO.getGrave().setFuneralId(returnDTO.getId());
+            if (returnDTO.getGrave().getDeceased() != null) {
+                for (DeceasedDTO deceasedDTO: returnDTO.getGrave().getDeceased()) {
+                    deceasedDTO.setGraveId(returnDTO.getGrave().getId());
+                }
+            }
+        }
+
+        return ResponseEntity.status(HttpStatus.OK).body(returnDTO);
     }
 
     @Override
@@ -104,16 +156,50 @@ public class FuneralServiceImpl implements FuneralService {
     }
 
     @Override
+    @Transactional
     public FuneralDTO saveAndReturnDTO(Funeral funeral) {
         Funeral savedFuneral = funeralRepository.save(funeral);
 
+        if (funeral.getGrave() != null) {
+            Grave grave = funeral.getGrave();
+            grave.setFuneral(funeral);
+            graveRepository.save(grave);
+            if (!Objects.equals(grave.getDeceased(), new HashSet<>())) {
+                for (Deceased deceased: grave.getDeceased()) {
+                    deceased.setGrave(grave);
+                    deceasedRepository.save(deceased);
+                }
+            }
+        }
+
         FuneralDTO returnDTO = funeralMapper.funeralToFuneralDTO(savedFuneral);
+        if (!Objects.equals(returnDTO.getGrave(), null)) {
+            returnDTO.getGrave().setFuneralId(returnDTO.getId());
+            if (!Objects.equals(returnDTO.getGrave().getDeceased(), new HashSet<>())) {
+                for (DeceasedDTO deceased: returnDTO.getGrave().getDeceased()) {
+                    deceased.setGraveId(returnDTO.getId());
+                }
+            }
+        }
 
         return returnDTO;
     }
 
     @Override
-    public void deleteFuneralById(Long id) {
-        funeralRepository.deleteById(id);
+    public ResponseEntity deleteFuneralById(Long id) {
+        Optional<Funeral> funeralOptional = funeralRepository.findById(id);
+
+        if (funeralOptional.isPresent()) {
+            Funeral funeral = funeralOptional.get();
+            if (funeral.getGrave() != null) {
+                funeral.getGrave().setFuneral(null);
+                funeral.setGrave(null);
+            }
+            funeralRepository.deleteById(id);
+
+            return ResponseEntity.status(HttpStatus.OK).body("Successfully deleted funeral with id: " + id.toString());
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Funeral not found for id: " + id.toString());
+        }
     }
 }
